requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.1
xgboost==2.1.3
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
joblib==1.4.2
matplotlib==3.9.3
seaborn==0.13.2
asyncpg==0.30.0
